requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# LinkedIn Automation - Approach 2 (Playwright)
fake-useragent>=1.4.0
numpy>=1.24.0

# LinkedIn Automation - Approach 3 (Third-Party)
apify-client>=1.6.0
httpx>=0.25.0

# LinkedIn Automation - Approach 1 (Official API)
# httpx already included above

# Shared utilities
pydantic>=2.5.0
